using flashdrive for project storage
creating python virtual environment in flashdrive are causing issues due to permission. This is the fix solution to start your project and use the flashdrive as a storage to all python libraries. Ordinary mounting from flashdrive and putting your projects in the flashdrive might cause permission issues and failed it to run.